fix(ui): Fix and cleanup related workflow runs
