--- frontend/src/components/executions/testdecode.ts
@@ -1,5 +1,0 @@
-console.log(
-  decodeURIComponent(
-    "sch-e9dd815289174fc3ad95261a12927ad2-2024-11-27T19%3A53%3A20Z"
-  )
-)
